refactor(pipeline): reorganize era-summary and tiles modules by domain
Move era-summary sync and validation modules from stages/ to era-summary/ directory, and tiles-related modules from stages/ to tiles/ directory, aligning the physical structure with domain and feature boundaries.
