+++ apps/pipeline/src/cli.ts
@@ -2,11 +2,11 @@ import path from 'node:path';
 import { listYears } from '@/commands/list-years.ts';
 import { showStatus } from '@/commands/status.ts';
 import { EXIT_CODES, PATHS } from '@/config.ts';
+import { syncEraSummaries } from '@/era-summary/sync.ts';
 import type { PipelineOptions } from '@/pipeline.ts';
 import { PipelineError, runPipeline } from '@/pipeline.ts';
 import type { PipelineLogger } from '@/shared/logger.ts';
 import { createLogger } from '@/shared/logger.ts';
-import { syncEraSummaries } from '@/stages/sync-era-summaries.ts';
 import { syncDescriptions } from '@/territory/sync.ts';
 import { validateAllDescriptions } from '@/territory/validate.ts';
