+++ apps/pipeline/src/era-summary/validate.ts
